Categorize Macro Playgorud by Macro Type
